Create new pages for quick view modal and block cart modal
