+++ tests/UI/campaigns/functional/FO/classic/02_headerAndFooter/01_checkLinksInHeader.ts
@@ -8,6 +8,8 @@ import {contactUsPage} from '@pages/FO/classic/contactUs';
 import {homePage} from '@pages/FO/classic/home';
 import {loginPage} from '@pages/FO/classic/login';
 import {myAccountPage} from '@pages/FO/classic/myAccount';
+import {blockCartModal} from '@pages/FO/classic/modal/blockCart';
+import {quickViewModal} from '@pages/FO/classic/modal/quickView';
 
 // Import data
 import Customers from '@data/demo/customers';
@@ -94,10 +96,11 @@ describe('FO - Header and Footer : Check links in header page', async () => {
 
     await loginPage.goToHomePage(page);
     // Add product to cart by quick view
-    await homePage.addProductToCartByQuickView(page, 1, 3);
+    await homePage.quickViewProduct(page, 1);
+    await quickViewModal.setQuantityAndAddToCart(page, 3);
 
     // Close block cart modal
-    const isQuickViewModalClosed = await homePage.closeBlockCartModal(page);
+    const isQuickViewModalClosed = await blockCartModal.closeBlockCartModal(page);
     expect(isQuickViewModalClosed).to.eq(true);
   });
