+++ tests/UI/campaigns/functional/FO/classic/09_productPage/01_quickView/01_addToCart.ts
@@ -14,6 +14,9 @@ import {ProductAttribute} from '@data/types/product';
 
 import {expect} from 'chai';
 import {BrowserContext, Page} from 'playwright';
+import {quickViewModal} from "@pages/FO/classic/modal/quickView";
+import {blockCartModal} from "@pages/FO/classic/modal/blockCart";
+import blockCart from "@pages/FO/hummingbird/modal/blockCart";
 
 const baseContext: string = 'functional_FO_classic_productPage_quickView_addToCart';
 
@@ -72,14 +75,17 @@ describe('FO - Product page - Quick view : Add to cart', async () => {
   it('should add first product to cart by quick view', async function () {
     await testContext.addContextItem(this, 'testIdentifier', 'addToCartByQuickView', baseContext);
 
-    const successMessage = await homePage.addProductToCartByQuickView(page, 1, 1);
+    await homePage.quickViewProduct(page, 1);
+    await quickViewModal.addToCartByQuickView(page);
+
+    const successMessage = await blockCartModal.getBlockCartModalTitle(page);
     expect(successMessage).to.contains(homePage.successAddToCartMessage);
   });
 
   it('should check product details from cart modal', async function () {
     await testContext.addContextItem(this, 'testIdentifier', 'checkProductDetailsInCartModal', baseContext);
 
-    const result = await homePage.getProductDetailsFromBlockCartModal(page);
+    const result = await blockCartModal.getProductDetailsFromBlockCartModal(page);
     await Promise.all([
       expect(result.name).to.equal(checkProductDetails.name),
       expect(result.price).to.equal(checkProductDetails.price),
@@ -90,7 +96,7 @@ describe('FO - Product page - Quick view : Add to cart', async () => {
       expect(result.totalTaxIncl).to.equal(checkProductDetails.totalTaxIncl),
     ]);
 
-    const productAttributesFromBlockCart = await homePage.getProductAttributesFromBlockCartModal(page);
+    const productAttributesFromBlockCart = await blockCartModal.getProductAttributesFromBlockCartModal(page);
     await Promise.all([
       expect(productAttributesFromBlockCart.length).to.equal(2),
       expect(productAttributesFromBlockCart[0].name).to.equal(checkProductDetailsProducts[0].name),
@@ -103,7 +109,7 @@ describe('FO - Product page - Quick view : Add to cart', async () => {
   it('should proceed to checkout and check the cart page', async function () {
     await testContext.addContextItem(this, 'testIdentifier', 'checkCartPage', baseContext);
 
-    await homePage.proceedToCheckout(page);
+    await blockCartModal.proceedToCheckout(page);
 
     const pageTitle = await cartPage.getPageTitle(page);
     expect(pageTitle).to.equal(cartPage.pageTitle);
