+++ tests/UI/campaigns/functional/FO/hummingbird/06_homePage/04_productQuickView.ts
@@ -10,6 +10,8 @@ import {installHummingbird, uninstallHummingbird} from '@commonTests/FO/hummingb
 // Import FO pages
 import homePage from '@pages/FO/hummingbird/home';
 import categoryPage from '@pages/FO/hummingbird/category';
+import quickViewModal from '@pages/FO/hummingbird/modal/quickView';
+import blockCartModal from '@pages/FO/hummingbird/modal/blockCart';
 
 // Import data
 import Products from '@data/demo/products';
@@ -83,7 +85,7 @@ describe('FO - Home Page : Product quick view', async () => {
       await homePage.goTo(page, global.FO.URL);
 
       const result = await homePage.isHomePage(page);
-      expect(result).to.eq(true);
+      expect(result).to.equal(true);
     });
 
     it(`should quick view the product '${Products.demo_6.name}'`, async function () {
@@ -91,8 +93,8 @@ describe('FO - Home Page : Product quick view', async () => {
 
       await homePage.quickViewProduct(page, 3);
 
-      const isModalVisible = await homePage.isQuickViewProductModalVisible(page);
-      expect(isModalVisible).to.eq(true);
+      const isModalVisible = await quickViewModal.isQuickViewProductModalVisible(page);
+      expect(isModalVisible).to.equal(true);
     });
 
     it('should check product information', async function () {
@@ -108,7 +110,7 @@ describe('FO - Home Page : Product quick view', async () => {
         expect(result.thumbImage).to.contains(Products.demo_6.thumbImage),
       ]);
 
-      const resultAttributes = await homePage.getSelectedAttributesFromQuickViewModal(page, defaultAttributes);
+      const resultAttributes = await quickViewModal.getSelectedAttributesFromQuickViewModal(page, defaultAttributes);
       expect(resultAttributes.length).to.be.equal(1);
       expect(resultAttributes[0].name).to.be.equal(defaultAttributes.name);
       expect(resultAttributes[0].value).to.be.equal(defaultAttributes.value);
@@ -117,9 +119,9 @@ describe('FO - Home Page : Product quick view', async () => {
     it('should change combination and check product information', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'changeCombination', baseContext);
 
-      await homePage.changeAttributes(page, attributes);
+      await quickViewModal.setAttributes(page, attributes);
 
-      const result = await homePage.getProductDetailsFromQuickViewModal(page);
+      const result = await quickViewModal.getProductDetailsFromQuickViewModal(page);
       await Promise.all([
         expect(result.name).to.equal(Products.demo_6.name),
         expect(result.price).to.equal(Products.demo_6.combinations[1].price),
@@ -128,7 +130,7 @@ describe('FO - Home Page : Product quick view', async () => {
         expect(result.thumbImage).to.contains(Products.demo_6.thumbImage),
       ]);
 
-      const resultAttributes = await homePage.getSelectedAttributesFromQuickViewModal(page, attributes);
+      const resultAttributes = await quickViewModal.getSelectedAttributesFromQuickViewModal(page, attributes);
       expect(resultAttributes.length).to.be.equal(1);
       expect(resultAttributes[0].name).to.be.equal(attributes.name);
       expect(resultAttributes[0].value).to.be.equal(attributes.value);
@@ -137,18 +139,18 @@ describe('FO - Home Page : Product quick view', async () => {
     it('should change the product quantity and click on add to cart', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'addProductToCart', baseContext);
 
-      await homePage.changeQuantity(page, attributesQty);
-      await homePage.addToCartByQuickView(page);
+      await quickViewModal.setQuantity(page, attributesQty);
+      await quickViewModal.addToCartByQuickView(page);
 
-      const isVisible = await homePage.isBlockCartModalVisible(page);
-      expect(isVisible).to.eq(true);
+      const isVisible = await blockCartModal.isBlockCartModalVisible(page);
+      expect(isVisible).to.equal(true);
     });
 
     it('should click on continue shopping and check that the modal is not visible', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'clickOnContinueShopping', baseContext);
 
-      const isNotVisible = await homePage.continueShopping(page);
-      expect(isNotVisible).to.eq(true);
+      const isNotVisible = await blockCartModal.continueShopping(page);
+      expect(isNotVisible).to.equal(true);
     });
   });
 
@@ -158,8 +160,8 @@ describe('FO - Home Page : Product quick view', async () => {
 
       await homePage.quickViewProduct(page, 6);
 
-      const isModalVisible = await homePage.isQuickViewProductModalVisible(page);
-      expect(isModalVisible).to.eq(true);
+      const isModalVisible = await quickViewModal.isQuickViewProductModalVisible(page);
+      expect(isModalVisible).to.equal(true);
     });
 
     it('should check product information', async function () {
@@ -179,18 +181,18 @@ describe('FO - Home Page : Product quick view', async () => {
     it('should change the product quantity and click on add to cart', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'addProductToCart2', baseContext);
 
-      await homePage.changeQuantity(page, attributesQty);
-      await homePage.addToCartByQuickView(page);
+      await quickViewModal.setQuantity(page, attributesQty);
+      await quickViewModal.addToCartByQuickView(page);
 
-      const isVisible = await homePage.isBlockCartModalVisible(page);
-      expect(isVisible).to.eq(true);
+      const isVisible = await blockCartModal.isBlockCartModalVisible(page);
+      expect(isVisible).to.equal(true);
     });
 
     it('should click on continue shopping and check that the modal is not visible', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'clickOnContinueShopping2', baseContext);
 
-      const isNotVisible = await homePage.continueShopping(page);
-      expect(isNotVisible).to.eq(true);
+      const isNotVisible = await blockCartModal.continueShopping(page);
+      expect(isNotVisible).to.equal(true);
     });
   });
 
@@ -201,7 +203,7 @@ describe('FO - Home Page : Product quick view', async () => {
       await homePage.clickOnAllProductsButton(page);
 
       const isCategoryPageVisible = await categoryPage.isCategoryPage(page);
-      expect(isCategoryPageVisible, 'Home category page was not opened').to.eq(true);
+      expect(isCategoryPageVisible, 'Home category page was not opened').to.equal(true);
     });
 
     it(`should go to the second page and quick view the product '${Products.demo_14.name}'`, async function () {
@@ -210,14 +212,14 @@ describe('FO - Home Page : Product quick view', async () => {
       await categoryPage.goToNextPage(page);
       await categoryPage.quickViewProduct(page, 7);
 
-      const isModalVisible = await categoryPage.isQuickViewProductModalVisible(page);
-      expect(isModalVisible).to.eq(true);
+      const isModalVisible = await quickViewModal.isQuickViewProductModalVisible(page);
+      expect(isModalVisible).to.equal(true);
     });
 
     it('should check product information', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'checkCustomizedProductInformation', baseContext);
 
-      const result = await homePage.getProductDetailsFromQuickViewModal(page);
+      const result = await quickViewModal.getProductDetailsFromQuickViewModal(page);
       await Promise.all([
         expect(result.name).to.equal(Products.demo_14.name),
         expect(result.price).to.equal(Products.demo_14.price),
@@ -231,15 +233,15 @@ describe('FO - Home Page : Product quick view', async () => {
     it('should check that \'Add to cart\' button is disabled', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'checkAddToCartButton', baseContext);
 
-      const isEnabled = await homePage.isAddToCartButtonEnabled(page);
-      expect(isEnabled, 'Add to cart button is not disabled').to.eq(false);
+      const isEnabled = await quickViewModal.isAddToCartButtonEnabled(page);
+      expect(isEnabled, 'Add to cart button is not disabled').to.equal(false);
     });
 
     it('should close the quick view modal', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'closeQuickOptionModal', baseContext);
 
-      const isQuickViewModalClosed = await homePage.closeQuickViewModal(page);
-      expect(isQuickViewModalClosed).to.eq(true);
+      const isQuickViewModalClosed = await quickViewModal.closeQuickViewModal(page);
+      expect(isQuickViewModalClosed).to.equal(true);
     });
   });
 
@@ -249,29 +251,29 @@ describe('FO - Home Page : Product quick view', async () => {
 
       await categoryPage.quickViewProduct(page, 8);
 
-      const isModalVisible = await categoryPage.isQuickViewProductModalVisible(page);
-      expect(isModalVisible).to.eq(true);
+      const isModalVisible = await quickViewModal.isQuickViewProductModalVisible(page);
+      expect(isModalVisible).to.equal(true);
     });
 
     it('should check that \'Add to cart\' button is disabled', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'checkAddToCartButton2', baseContext);
 
-      const isEnabled = await homePage.isAddToCartButtonEnabled(page);
-      expect(isEnabled, 'Add to cart button is not disabled').to.eq(false);
+      const isEnabled = await quickViewModal.isAddToCartButtonEnabled(page);
+      expect(isEnabled, 'Add to cart button is not disabled').to.equal(false);
     });
 
     it('should check the product availability', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'checkProductAvailability', baseContext);
 
-      const availability = await homePage.getProductAvailabilityText(page);
+      const availability = await quickViewModal.getProductAvailabilityText(page);
       expect(availability).to.contains('Out-of-Stock');
     });
 
     it('should close the quick view modal', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'closeQuickOptionModal2', baseContext);
 
-      const isQuickViewModalClosed = await homePage.closeQuickViewModal(page);
-      expect(isQuickViewModalClosed).to.eq(true);
+      const isQuickViewModalClosed = await quickViewModal.closeQuickViewModal(page);
+      expect(isQuickViewModalClosed).to.equal(true);
     });
   });
