+++ tests/UI/campaigns/functional/FO/hummingbird/09_productPage/01_quickView/01_addToCart.ts
@@ -10,6 +10,8 @@ import {installHummingbird, uninstallHummingbird} from '@commonTests/FO/hummingb
 import homePage from '@pages/FO/hummingbird/home';
 import cartPage from '@pages/FO/hummingbird/cart';
 import searchResultsPage from '@pages/FO/hummingbird/searchResults';
+import quickViewModal from '@pages/FO/hummingbird/modal/quickView';
+import blockCartModal from '@pages/FO/hummingbird/modal/blockCart';
 
 // Import data
 import Products from '@data/demo/products';
@@ -85,14 +87,16 @@ describe('FO - Product page - Quick view : Add to cart', async () => {
     it('should add first product to cart by quick view', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'addToCartByQuickView', baseContext);
 
-      const successMessage = await homePage.addProductToCartByQuickView(page, 1, 1);
+      await homePage.quickViewProduct(page, 1);
+
+      const successMessage = await quickViewModal.addToCartByQuickView(page);
       expect(successMessage).to.contains(homePage.successAddToCartMessage);
     });
 
     it('should check product details from cart modal', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'checkProductDetailsInCartModal', baseContext);
 
-      const result = await homePage.getProductDetailsFromBlockCartModal(page);
+      const result = await blockCartModal.getProductDetailsFromBlockCartModal(page);
       await Promise.all([
         expect(result.name).to.equal(checkProductDetails.name),
         expect(result.price).to.equal(checkProductDetails.price),
@@ -103,7 +107,7 @@ describe('FO - Product page - Quick view : Add to cart', async () => {
         expect(result.totalTaxIncl).to.equal(checkProductDetails.totalTaxIncl),
       ]);
 
-      const productAttributesFromBlockCart = await homePage.getProductAttributesFromBlockCartModal(page);
+      const productAttributesFromBlockCart = await blockCartModal.getProductAttributesFromBlockCartModal(page);
       await Promise.all([
         expect(productAttributesFromBlockCart.length).to.equal(2),
         expect(productAttributesFromBlockCart[0].name).to.equal(checkProductDetailsProducts[0].name),
@@ -116,7 +120,7 @@ describe('FO - Product page - Quick view : Add to cart', async () => {
     it('should proceed to checkout and check the cart page', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'checkCartPage', baseContext);
 
-      await homePage.proceedToCheckout(page);
+      await blockCartModal.proceedToCheckout(page);
 
       const pageTitle = await cartPage.getPageTitle(page);
       expect(pageTitle).to.equal(cartPage.pageTitle);
@@ -169,7 +173,7 @@ describe('FO - Product page - Quick view : Add to cart', async () => {
 
       await searchResultsPage.quickViewProduct(page, 1);
 
-      const isDisabled = await homePage.isAddToCartButtonDisabled(page);
+      const isDisabled = await quickViewModal.isAddToCartButtonDisabled(page);
       expect(isDisabled).to.eq(true);
     });
   });
