+++ tests/UI/campaigns/functional/FO/hummingbird/09_productPage/01_quickView/03_shareLinks.ts
@@ -1,0 +1,72 @@
+// Import utils
+import helper from '@utils/helpers';
+import testContext from '@utils/testContext';
+import files from '@utils/files';
+
+// Import common tests
+import {installHummingbird, uninstallHummingbird} from '@commonTests/FO/hummingbird';
+
+// Import pages
+import homePage from '@pages/FO/hummingbird/home';
+import quickViewModal from '@pages/FO/hummingbird/modal/quickView';
+
+import {expect} from 'chai';
+import type {BrowserContext, Page} from 'playwright';
+
+const baseContext: string = 'functional_FO_hummingbird_productPage_quickView_changeQuantity';
+
+/*
+Pre-condition:
+- Install hummingbird theme
+Scenario:
+- Go to FO
+- Quick view third product
+
+Post-condition:
+- Uninstall hummingbird theme
+ */
+describe('FO - Product page - Quick view : Share links', async () => {
+  let browserContext: BrowserContext;
+  let page: Page;
+
+  // Pre-condition : Install Hummingbird
+  installHummingbird(`${baseContext}_preTest`);
+
+  // before and after functions
+  before(async function () {
+    browserContext = await helper.createBrowserContext(this.browser);
+    page = await helper.newTab(browserContext);
+  });
+
+  after(async () => {
+    await helper.closeBrowserContext(browserContext);
+    await files.deleteFile('../../admin-dev/hummingbird.zip');
+  });
+
+  describe('Check share links', async () => {
+    it('should go to FO home page', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'goToFo', baseContext);
+
+      await homePage.goToFo(page);
+
+      const isHomePage = await homePage.isHomePage(page);
+      expect(isHomePage).to.eq(true);
+    });
+
+    it('should quick view the third product', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'quickView', baseContext);
+
+      await homePage.quickViewProduct(page, 3);
+
+      const isModalVisible = await quickViewModal.isQuickViewProductModalVisible(page);
+      expect(isModalVisible).to.eq(true);
+    });
+
+    it('should check the first share link \'F\'', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'checkFacebookLink', baseContext);
+    });
+  });
+
+  // Post-condition : Uninstall Hummingbird
+  uninstallHummingbird(`${baseContext}_postTest`);
+});
